registered models with the admin for userprofiles and wasteprocessors
